backend/api/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from .views import (CustonUserViewSet, IngredientViewSet, RecipeViewSet,
                    TagViewSet)

app_name = 'api'

router = DefaultRouter

router.register('users', CustonUserViewSet, basename='users')
router.register('ingredients', IngredientViewSet, basename='ingredients')
router.register('recipes', RecipeViewSet, basename='recipes')
router.register('tags', TagViewSet, basename='tags')

urlpatterns = [path('', include(router.urls))]
